refactor(game): move game-state-type into lib/core (T26)
Relocate the game state type definitions from packages/game/types/
to packages/game/lib/core/ as part of the layered architecture
restructuring. All nine importers updated to the new path. The
remaining types/env.d.ts (ambient __APP_VERSION__ declaration) is
unrelated to layer rules and stays in place.
