+++ packages/game/lib/update-enemies.ts
@@ -1,4 +1,4 @@
-import type { GameState } from "../types/game-state-type.ts";
+import type { GameState } from "./core/types.ts";
 import { moveEnemyTowardPlayer } from "./move-enemy-toward-player.js";
 
 /**
